HomeWork_18_Lite/requirements.txt
python-telegram-bot
python-telegram-bot[job-queue]
python-dotenv
openai
requests
aiohttp